azure-durable-functions/pyproject.toml
[project]
name = "workflow-orchestration-adf"
version = "0.1.0"
description = "Azure Durable Functions implementation of invoice processing workflow (part of workflow-orchestration-poc monorepo)"
readme = "README.md"
requires-python = "==3.11.12"
dependencies = [
    "azure-functions-durable==1.4.0",
    "azure-functions==1.24.0",
    "azure-storage-blob==12.19.0",
    "pillow==10.0.0",
    "pymupdf==1.24.0",
]
